requirements.txt
python-telegram-bot[job-queue]==20.7
Flask==3.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
werkzeug==3.0.1
gunicorn==21.2.0
waitress==2.1.2
